Update Makefile with current URL for sounds
The Asterisk sounds repository has moved to http://downloads.asterisk.org.  This causes the build to fail, because when the script attempts to download the sounds archive, it gets a 302 HTTP redirect and downloads a HTML listing of the repo directory instead.

This PR provides the correct URL in SOUNDS_URL in the Makefile.